Reorganized the file structure of the project to separate each game state into its own file.
