+++ src/screen/js/game_states/end.js
@@ -1,0 +1,12 @@
+// End
+module.exports = function() {};
+module.exports.prototype = {
+	preload: function() {
+	},
+	create: function() {
+	},
+	update: function() {	
+	},
+	render: function() {
+	}
+};
