+++ src/screen/js/game_states/minigame_info.js
@@ -1,0 +1,16 @@
+// Minigame Info
+module.exports = function() {};
+module.exports.prototype = {
+	preload: function() {
+		
+	},
+	create: function() {
+		
+	},
+	update: function() {
+		
+	},
+	render: function() {
+		
+	}
+};
