+++ src/screen/js/game_states/play.js
@@ -1,0 +1,12 @@
+// Play state
+module.exports = function() {};
+module.exports.prototype = {
+	preload: function() {
+	},
+	create: function() {
+	},
+	update: function() {	
+	},
+	render: function() {
+	}
+};
