+++ src/screen/screen.js
@@ -1,2 +1,2 @@
 require('./js/phaser.min.js');
-require('./js/main.js');
+require('./js/app.js');
